Convert to maven compiling and fixed this ghay code
- Convert to maven instead of weird IntelliJ
- Added fallback to protocollib in case it isn't installed
- Made the plugin disable itself on database connection failure
- Adds 1.13.2 support
- Added support to not need to download spigot to compile this
- Supresses some deprecation & usage warnings
- Fixes ghay code
