--- .idea/artifacts/FreedomCore_jar.xml
@@ -1,9 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="FreedomCore:jar">
-    <output-path>$PROJECT_DIR$/out/artifacts/FreedomCore_jar</output-path>
-    <root id="archive" name="FreedomCore.jar">
-      <element id="module-output" name="FreedomCore" />
-      <element id="extracted-dir" path="E:/mongo-java-driver-3.9.1.jar" path-in-jar="/" />
-    </root>
-  </artifact>
-</component>
