--- .idea/beamePluginConfig1.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="PluginConfig">
-    <option name="_beameInstaCmdLocation" value="" />
-    <option name="_nodeJsCmdLocation" value="C:\Program Files\nodejs\node.exe" />
-  </component>
-</project>
