--- .idea/libraries/ProtocolLib.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="ProtocolLib">
-    <CLASSES>
-      <root url="jar://E:/ProtocolLib.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
