--- .idea/libraries/mongo_java_driver_3_9_1.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="mongo-java-driver-3.9.1">
-    <CLASSES>
-      <root url="jar://E:/mongo-java-driver-3.9.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
