--- .idea/libraries/spigot_1_12_2_R0_1_SNAPSHOT_b1498.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="spigot-1.12.2-R0.1-SNAPSHOT-b1498">
-    <CLASSES>
-      <root url="jar://F:/spigot-1.12.2-R0.1-SNAPSHOT-b1498.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://F:/spigot-1.12.2-R0.1-SNAPSHOT-b1498.jar!/" />
-    </SOURCES>
-  </library>
-</component>
