--- FreedomCore.iml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="spigot-1.12.2-R0.1-SNAPSHOT-b1498" level="project" />
-    <orderEntry type="library" name="mongo-java-driver-3.9.1" level="project" />
-    <orderEntry type="library" name="ProtocolLib" level="project" />
-  </component>
-</module>
